rebase package.json
